requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
pywin32
pillow>=8.0.0
requests>=2.25.0
pyserial>=3.5
# pybluez>=0.23  # optionnel: scan Bluetooth radio (pip install pybluez)
